Lab2 Vector and Stack

+++ LibraryCPP/stack.h
@@ -3,7 +3,7 @@
 
 // Stack
 // Stores integer values inside
-typedef int Data;
+typedef size_t Data;
 
 struct Stack;
